# openapi/app.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Title",
    "description": "Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/room/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "organization": {
                    "type": "string"
                  },
                  "discussionFlag": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {

        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy",
          "uri": "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:${AWS::AccountId}:function:daochat-Chat-Api/invocations",
          "credentials": "arn:aws:iam::012688626467:role/apigwrole"
        },
        "security": [
          {
            "CognitoUserPool": []
          }
        ]
      },
      "options": {
        "responses": {
          "200": {
            "description": "default",
            "headers": {
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "mock",
          "requestTemplates": {
            "application/json": "{'statusCode' : 200}"
          },
          "responses": {
            "default": {
              "statusCode": 200,
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Headers": "'*'",
                "method.response.header.Access-Control-Allow-Methods": "'OPTIONS,POST,GET'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          }
        }
      }
    },
    "/room/list": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "schema": {
              "type": "string"
            },
            "name": "id",
            "required": true
          }
        ],
        "responses": {

        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy",
          "uri": "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:${AWS::AccountId}:function:daochat-Chat-Api/invocations",
          "credentials": "arn:aws:iam::012688626467:role/apigwrole"
        },
        "security": [
          {
            "CognitoUserPool": []
          }
        ]
      },
      "options": {
        "responses": {
          "200": {
            "description": "default",
            "headers": {
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "x-amazon-apigateway-integration": {
            "httpMethod": "POST",
            "passthroughBehavior": "when_no_match",
            "type": "aws_proxy",
            "uri": "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:${AWS::AccountId}:function:daochat-Admin-Api/invocations",
            "credentials": "arn:aws:iam::012688626467:role/apigwrole"
          },
          "security": [
            {
              "CognitoUserPool": []
            }
          ]
        },
        "x-amazon-apigateway-integration": {
          "type": "mock",
          "requestTemplates": {
            "application/json": "{'statusCode' : 200}"
          },
          "responses": {
            "default": {
              "statusCode": 200,
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Headers": "'*'",
                "method.response.header.Access-Control-Allow-Methods": "'OPTIONS,POST,GET'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          }
        }
      }
    },
    "/room": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {

        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy",
          "uri": "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:${AWS::AccountId}:function:daochat-Chat-Api/invocations",
          "credentials": "arn:aws:iam::012688626467:role/apigwrole"
        },
        "security": [
          {
            "CognitoUserPool": []
          }
        ]
      },
      "options": {
        "responses": {
          "200": {
            "description": "default",
            "headers": {
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "mock",
          "requestTemplates": {
            "application/json": "{'statusCode' : 200}"
          },
          "responses": {
            "default": {
              "statusCode": 200,
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Headers": "'*'",
                "method.response.header.Access-Control-Allow-Methods": "'OPTIONS,POST,GET'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          }
        }
      }
    },
    "/room/chat": {
      "get": {
        "parameters": [{
          "in": "query",
          "schema": {
            "type": "string"
          },
          "required": true,
          "name": "id"
        }],
        "responses": {

        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy",
          "uri": "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:${AWS::AccountId}:function:daochat-Chat-Api/invocations",
          "credentials": "arn:aws:iam::012688626467:role/apigwrole"
        },
        "security": [
          {
            "CognitoUserPool": []
          }
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {

        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy",
          "uri": "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:${AWS::AccountId}:function:daochat-Chat-Api/invocations",
          "credentials": "arn:aws:iam::012688626467:role/apigwrole"
        },
        "security": [
          {
            "CognitoUserPool": []
          }
        ]
      },
      "options": {
        "responses": {
          "200": {
            "description": "default",
            "headers": {
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "mock",
          "requestTemplates": {
            "application/json": "{'statusCode' : 200}"
          },
          "responses": {
            "default": {
              "statusCode": 200,
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Headers": "'*'",
                "method.response.header.Access-Control-Allow-Methods": "'OPTIONS,POST,GET'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          }
        }
      }
    },
...
    "/room/vote/push": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "schema": {
              "type": "string"
            },
            "name": "id",
            "required": true
          }
        ],
        "responses": {

        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy",
          "uri": "arn:aws:apigateway:us-east-1:lambda:path/2015-03-31/functions/arn:aws:lambda:us-east-1:${AWS::AccountId}:function:daochat-Chat-Api/invocations",
          "credentials": "arn:aws:iam::012688626467:role/apigwrole"
        },
        "security": [
          {
            "CognitoUserPool": []
          }
        ]
      },
      "options": {
        "responses": {
          "200": {
            "description": "default",
            "headers": {
              "Access-Control-Allow-Headers": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Methods": {
                "schema": {
                  "type": "string"
                }
              },
              "Access-Control-Allow-Origin": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "type": "mock",
          "requestTemplates": {
            "application/json": "{'statusCode' : 200}"
          },
          "responses": {
            "default": {
              "statusCode": 200,
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Headers": "'*'",
                "method.response.header.Access-Control-Allow-Methods": "'OPTIONS,POST,GET'",
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CognitoUserPool": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "x-amazon-apigateway-authtype": "cognito_user_pools",
        "x-amazon-apigateway-authorizer": {
          "type": "cognito_user_pools",
          "providerARNs": [
            "arn:aws:cognito-idp:us-east-1:${AWS::AccountId}:userpool/us-east-1_pgMduxM1X"
          ]
        }
      }
    }
  }
}